requirements.txt
python-dotenv~=1.0.1
openai ~=1.14.3
duckduckgo_search
colorama
openpyxl~=3.1.2
pandas~=2.2.2


